Abstract class for local search.
